fix: Run textlint on files with any extension
Success or failure of textlint depends on the extension of the target
file. That is, the target file must have .md or .text extension.
However, the file generated by the `mktemp` command cannot run textlint,
even if the file type is markdown or text, because the extension is a
random string.